# reshapr-public-openapi-v0.1.yaml
...
openapi: 3.0.2
info:
  title: reShapr Public API v0.1
  version: 0.0.1
  description: reShapr control plane REST API
  contact:
    name: Laurent Broudoux
    url: https://github.com/lbroudoux
    email: laurent.broudoux@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
  /v1/artifacts:
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  description: The specification file to import as an artifact
                  type: string
                  format: binary
                mainArtifact:
                  description: Whether this artifact is a primary one (default is true)
                  type: boolean
                  default: true
                serviceName:
                  description: Name to assign to the imported Service (overrides the one from the specification)
                  type: string
                serviceVersion:
                  description: Version to assign to the imported Service (overrides the one from the specification)
                  type: string
                includedOperations:
                  description: List of operation names to include from the specification
                  type: array
                  items:
                    type: string
                excludedOperations:
                  description: List of operation names to exclude from the specification
                  type: array
                  items:
                    type: string
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  description: The remote URL of the specification to import as an artifact
                  type: string
                mainArtifact:
                  description: Whether this artifact is a primary one (default is true)
                  type: boolean
                  default: true
                secretName:
                  description: Name of the Secret to use for authenticating the remote URL fetch
                  type: string
                serviceName:
                  description: Name to assign to the imported Service (overrides the one from the specification)
                  type: string
                serviceVersion:
                  description: Version to assign to the imported Service (overrides the one from the specification)
                  type: string
                includedOperations:
                  description: List of operation names to include from the specification
                  type: array
                  items:
                    type: string
                excludedOperations:
                  description: List of operation names to exclude from the specification
                  type: array
                  items:
                    type: string
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
          description: Artifact imported
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: >-
            The artifact could not be imported. The response body contains a message describing the
            validation error (e.g. invalid or malformed specification file, unresolvable remote URL).
      security:
      - jwt-bearer: []
      operationId: importArtifact
  /v1/artifacts/attach:
    post:
      requestBody:
        content:
          multipart/form-data: { }
          application/x-www-form-urlencoded: { }
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
          description: Artifact attache
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: >-
            The artifact could not be attached. The response body contains a message describing the
            validation error (e.g. invalid or malformed artifact file, unresolvable remote URL).
      security:
        - jwt-bearer: [ ]
      operationId: attachArtifact
  /v1/artifacts/{id}:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
          description: The artifact with matching identifier
        "404":
          description: Artifact with this id is unknown
      security:
        - jwt-bearer: [ ]
      operationId: getArtifact
    delete:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactDeletionImpact'
          description: >-
            Artifact deleted. The response describes the impact of the deletion on the ConfigurationPlans
            of the service that referenced the artifact by name in their includedArtifacts selection.
        "404":
          description: Artifact with this id is unknown
      security:
        - jwt-bearer: [ ]
      operationId: deleteArtifact
    parameters:
      - name: id
        description: Unique identifier of the Artifact to manage
        schema:
          type: string
        in: path
        required: true
  /v1/artifacts/{id}/deletion-impact:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactDeletionImpact'
          description: >-
            Non-mutating preview of the impact of deleting the artifact, to back a confirmation prompt.
        "404":
          description: Artifact with this id is unknown
      security:
        - jwt-bearer: [ ]
      operationId: getArtifactDeletionImpact
    parameters:
      - name: id
        description: Unique identifier of the Artifact to manage
        schema:
          type: string
        in: path
        required: true
  /v1/artifacts/service/{serviceId}:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artifact'
          description: Get the list of service artifacts
      operationId: getArtifactsByServiceId
    parameters:
      - name: serviceId
        description: Unique identifier of Service
        schema:
          type: string
        in: path
        required: true
...
components:
  schemas:
    Error:
      description: A structured error payload returned when a request cannot be fulfilled
      required:
      - message
      type: object
      properties:
        message:
          description: A human-readable description of the error
          type: string
    Artifact:
      description: An Artifact in reShapr control plane
      required:
      - id
      - organizationId
      - name
      - content
      - mainArtifact
      - type
      - sourceArtifact
      type: object
      properties:
        id:
          description: Artifact unique identifier
          type: string
        organizationId:
          description: Identifier of the owning organization
          type: string
        name:
          description: Artifact name
          type: string
        content:
          description: Artifact content
          type: string
        path:
          description: Relative path from root Artifact
          type: string
        mainArtifact:
          description: Whether this artifact is a primary one
          type: boolean
        type:
          $ref: '#/components/schemas/ArtifactType'
          description: The type of this Artifact
        sourceArtifact:
          description: Name or URL of original artifact file
          type: string
        capabilities:
          description: >-
            Names of the capabilities declared by a custom artifact (prompt names, resource/
            resourceTemplate uris, custom tool names or filters).
          type: array
          items:
            type: string
    ArtifactDeletionImpact:
      description: >-
        Impact of deleting an artifact on the ConfigurationPlans of its service. Returned as a preview by
        GET /v1/artifacts/{id}/deletion-impact and as the applied result by DELETE /v1/artifacts/{id}.
      required:
        - artifactId
        - artifactName
        - mainArtifact
        - impactedPlans
      type: object
      properties:
        artifactId:
          description: Identifier of the artifact
          type: string
        artifactName:
          description: Name of the artifact
          type: string
        mainArtifact:
          description: Whether the artifact is the service main artifact
          type: boolean
        impactedPlans:
          description: ConfigurationPlans referencing the artifact by name in their includedArtifacts
          type: array
          items:
            $ref: '#/components/schemas/ImpactedConfigurationPlan'
    ImpactedConfigurationPlan:
      description: A ConfigurationPlan impacted by an artifact deletion
      required:
        - id
        - name
        - fallsBackToAll
      type: object
      properties:
        id:
          description: Identifier of the ConfigurationPlan
          type: string
        name:
          description: Name of the ConfigurationPlan
          type: string
        fallsBackToAll:
          description: >-
            Whether removing the artifact empties the plan selection, making it fall back to all attached
            artifacts of the service
          type: boolean
    ArtifactType:
      description: The supported type of Artifact
      enum:
      - JSON_SCHEMA
      - OPEN_API_SPEC
      - GRAPHQL_SCHEMA
      - PROTOBUF_SCHEMA
      - PROTOBUF_DESCRIPTOR
      - JSON_FRAGMENT
      - RESHAPR_PROMPTS
      - RESHAPR_CUSTOM_TOOLS
      - RESHAPR_RESOURCES
      type: string
...
  securitySchemes:
    jwt-bearer:
      scheme: bearer
      type: http
      description: JWT Bearer acquired from reShapr Authentication API /auth endpoint